refactor(content): WKC layout — Phase 8 (garbage collection)
Phase 8 of the WKC migration. Moves the GC subsystem out of the legacy
src/service/ tree into the WKC layout.

Moved:

- service/garbage-collection/GarbageCollectionManager.ts ->
  logic/garbage-collection/component.ts
- logic/delete-unreferenced-files.ts ->
  logic/garbage-collection/delete-unreferenced-files.ts (sibling helper
  in the same component folder; both are GC operations on different
  schedules)

Type extraction:

- GCStaleProfilesResult and SweepResult moved to types.ts.

Wiring:

- All importers swept via sed:
  service/garbage-collection/GarbageCollectionManager -> logic/garbage-collection
  logic/delete-unreferenced-files -> logic/garbage-collection
- src/service/garbage-collection/ deleted entirely.
- src/service/ now contains only time/ (a tiny TimeSorting helper).

Behavior preserved:

- GarbageCollectionManager class shape unchanged (start/stop/performSweep
  semantics, sweep interval, profile-cleanup limits, transaction
  boundaries inside gcStaleProfiles).
- deleteUnreferencedFiles entry point and bloom-filter-based deletion
  flow unchanged; still exported for entrypoints/run-maintenance.ts.

tsc clean, unit suite green (213 tests, no diff).
