+++ content/src/components.ts
@@ -38,7 +38,7 @@ import { createSnapshotGenerator } from './ports/snapshotGenerator'
 import { createSnapshotStorage } from './ports/snapshotStorage'
 import { createSynchronizationState } from './adapters/synchronization-state'
 import { createSystemProperties } from './adapters/system-properties'
-import { GarbageCollectionManager } from './service/garbage-collection/GarbageCollectionManager'
+import { GarbageCollectionManager } from './logic/garbage-collection'
 import { PointerManager } from './logic/pointer-manager'
 import { ChallengeSupervisor } from './logic/challenge-supervisor'
 import { createContentCluster } from './logic/peer-cluster'
