+++ content/src/entrypoints/run-maintenance.ts
@@ -5,7 +5,7 @@ import { createLogComponent } from '@well-known-components/logger'
 import { createTestMetricsComponent } from '@dcl/metrics'
 import path from 'path'
 import { EnvironmentBuilder, EnvironmentConfig } from '../Environment'
-import { deleteUnreferencedFiles } from '../logic/delete-unreferenced-files'
+import { deleteUnreferencedFiles } from '../logic/garbage-collection'
 import { metricsDeclaration } from '../metrics'
 import { migrateContentFolderStructure } from '../migrations/ContentFolderMigrationManager'
 import { createMigrationExecutor } from '../migrations/migration-executor'
