+++ content/src/logic/garbage-collection/index.ts
@@ -1,0 +1,3 @@
+export { GarbageCollectionManager } from './component'
+export { deleteUnreferencedFiles } from './delete-unreferenced-files'
+export type { GCStaleProfilesResult, SweepResult } from './types'
