+++ content/src/logic/garbage-collection/types.ts
@@ -1,0 +1,10 @@
+export type GCStaleProfilesResult = {
+  deletedHashes: Set<string>
+  deletedDeployments: Set<string>
+}
+
+export type SweepResult = {
+  gcProfileActiveEntitiesResult?: Set<string>
+  gcUnusedHashResult?: Set<string>
+  gcStaleProfilesResult?: GCStaleProfilesResult
+}
