+++ content/test/integration/logic/delete-unreferenced-files.spec.ts
@@ -4,7 +4,7 @@ import { mkdirSync, mkdtempSync, rmSync } from 'fs'
 import os from 'os'
 import path from 'path'
 import { EnvironmentConfig } from '../../../src/Environment'
-import { deleteUnreferencedFiles } from '../../../src/logic/delete-unreferenced-files'
+import { deleteUnreferencedFiles } from '../../../src/logic/garbage-collection'
 import { MS_PER_DAY } from '../../../src/logic/time-range'
 import { hashFiles } from '../../../src/logic/deployment-service'
 import { AppComponents } from '../../../src/types'
